api/pyproject.toml
[project]
name = "puzzle-image-api"
version = "0.1.0"
description = "FastAPI backend for puzzle-image"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pillow>=10.3.0",
    "pydantic>=2.7.0",
    "sqlmodel>=0.0.16",
    "sqlalchemy>=2.0.30",
    "boto3>=1.34.0",
    "redis>=5.0.4",
    # Optional AI providers (uncomment when used)
    # "google-cloud-vision>=3.7.4",
    # "botocore>=1.34.0"
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["**/*.py"]

[tool.uvicorn]
reload = true
host = "0.0.0.0"
port = 8000
